requirements.txt
httpx>=0.27.0
pydantic>=2.6.0
tenacity>=8.2.3
APScheduler>=3.10.4
python-dotenv>=1.0.1
typer[all]>=0.12.3
orjson>=3.10.7
rich>=13.7.1
tzdata>=2024.1
